requirements.txt
absl-py==0.12.0
astor==0.8.1
atomicwrites==1.4.0
attrs==20.3.0
cached-property==1.5.2
category-encoders==2.2.2
certifi==2020.12.5
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
copulas==0.5.0
ctgan==0.4.1
cycler==0.10.0
decorator==4.4.2
deepecho==0.1.4
Faker==4.14.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
graphviz==0.16
grpcio==1.36.1
h5py==2.10.0
imageio==2.9.0
importlib-metadata==3.3.0
iniconfig==1.1.1
ipython==5.8.0
ipython-genutils==0.2.0
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1603996595157/work
Lasagne==0.1
lightgbm==3.1.1
llvmlite==0.35.0
Markdown==3.3.4
matplotlib @ file:///C:/ci/matplotlib-base_1592846084747/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numba==0.52.0
numpy==1.20.1
olefile==0.46
opacus==0.11.0
opt-einsum==3.3.0
packaging==20.8
pandas @ file:///C:/ci/pandas_1602088205210/work
patsy==0.5.1
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
pluggy==0.13.1
pomegranate==0.14.1
prompt-toolkit==1.0.15
protobuf==3.15.6
py==1.10.0
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
pyvacy==0.0.32
PyYAML==5.3.1
rdt==0.4.1
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
sdmetrics==0.2.0
sdv==0.0.0
seaborn==0.11.1
simplegeneric==0.8.1
six==1.15.0
sktime==0.5.3
statsmodels==0.12.2
tabgan==1.0.3
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
text-unidecode==1.3
Theano==1.0.5
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
